Optimize distinguishable_colors
The most important aspect of this change is storing candidates
in Lab format, which is the "intrinsic" format used in `colordiff`
(with DE_2000). If we're seeking n colors and have N candidates
(with N >> n), the algorithm makes O(n*N) difference computations;
having all of these start out in Lab format reduces the number
of RGB->Lab conversions (which are expensive, due to the 2.4 power)
to O(n).

There are a couple of other tweaks, including making the default
seed a concrete type and speeding up some 7th power computations.