The big rename, plus color("r") -> parse(Color, "r")
This updates to the new names in ColorTypes. Because of potentially
ambiguous meaning to `color(arg)` (which could be passed a string
or a Opaque/TransparentColor; but what should it do for the latter?),
I narrowed and aligned it with recent changes in Julia to become a
method for `parse`. In digging into this I also uncovered and fixed a
bug in the parsing of "hsl(180,40%,30%)". I also made it more
CSS-compliant by clamping the RGB values to [0,1] before returning
the result. Finally, I added tests for the new parse behavior.