Add methods for conversion with the white point
`Lab` and `Luv` have conversion method with the white point already.
However, the Lab/Luv-based color types, i.e. LCHab, LCHuv, DIN99 and DIN99o, do not.
This comit add the interface to support the white point for ALL conversions, including from RGB to HSV and so on.
In the conversion which do not depend on the white point, the fallbacks omit the white point.