Successfully precompile `_parse_colorant`
This is an expensive method to infer (>50ms on my machine) and it
has never precompiled successfully. It turns out that
calling it during module definition makes it non-precompilable.
I am guessing that this is because there end up being backedges
to something in the temporary process that does the precompilation
that is no longer available, perhaps gensyms in macro expansion?
In any case, avoiding the call works around the problem.