Improve speed of `pow5_12(::Float64)` and accuracy of `pow5_12(::Float32)`
`pow5_12(::Float64)` is used in the `XYZ{Float64}`-->`RGB` conversion.
This ensures sufficient accuracy (~3 ULP) in the range of [0.003, 1],
which is typically required for the conversion.

Previously, even conversions from `XYZ{::Float32}` were calculated with `Float64`.
This changes the conversions so that `pow5_12(::Float32)` is used.
However, `pow5_12(::Float32)` still be calculated within `Float64` for accuracy.

`@inline` could be a workaround for the problem with precompilation.