Optimize `ColorTypes.hue(c)`
The conventional version using `atand` has a larger numerical error
because `rad2deg` is applied to the `atan` result.
This improves the accuracy and speed by making the internal representation
conform to degrees and by simplifying the handling of exceptional cases.

This change should be made in ColorTypes.jl,
but for the convenience of development, it is implemented once in Colors.jl.