remove invalid fastmath usage in atan360
The functions need to be able to deal with infinities and NaNs but fastmath assumes that arguments and results are not so.

On 1.12:

```
julia> Colors.atan360(-Inf, Inf)
NaN
```

with fastmath removed:

```
julia> Colors.atan360(-Inf, Inf)
315.0
```

I haven't looked into any of the other fastmath usage in this package.